ClusterRules: combine again filtering tests
